Move axios configuration under api/axios
